+++ frontend/src/router/Router.tsx
@@ -1,7 +1,7 @@
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import App from '../App';
 import NotFound from '../pages/NotFound';
-import Home from '../components/home/Home';
+import Home from '../pages/Home';
 
 const routes = createBrowserRouter([
   {
